requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langchain-core
langchain-google-genai
langgraph
presidio-analyzer
presidio-anonymizer
spacy>=3.8.0
